back/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask-Cors==4.0.0
flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pip==24.0
setuptools==65.5.1
werkzeug==3.0.2
wheel==0.38.4
tika~=2.6.0
requests~=2.23.0
beautifulsoup4~=4.8.2
parse_hh_data==0.1.14
Pillow